requirements.txt
discord.py
discord-pretty-help
psycopg2-binary